alert_system/requirements.txt
mysql-connector-python==9.1.0
confluent-kafka==1.9.2